Extract iOS analytics transport
